tests/Document-getElementsByTagNameNS.html

<!DOCTYPE html>
<title>Document.getElementsByTagNameNS</title>
<link rel="author" title="Ms2ger" href="mailto:ms2ger@gmail.com">
<link rel="help" href="../web-dom-core#dom-document-getelementsbytagnamens">
<script src="testharness.js"></script>
<body>
<div id="log"></div>
<script>
test(function() {
  assert_true(document.getElementsByTagNameNS("http://www.w3.org/1999/xhtml", "html") instanceof NodeList, "NodeList")
  assert_false(document.getElementsByTagNameNS("http://www.w3.org/1999/xhtml", "html") instanceof HTMLCollection, "HTMLCollection")
  assert_true(document.getElementsByTagNameNS("http://www.w3.org/1999/xhtml", "html") !== document.getElementsByTagNameNS("http://www.w3.org/1999/xhtml", "html"), "no caching")
})
</script>
